// frontend/src/app/components/model/ambulance.model.ts
export interface Ambulance {
  ambulanceId: number;
  additional_notes: string;
  area: string;
  block_no: string;
  contact_name: string;
  contact_phone: string;
  emergency_ambulance: string;
  freezing_ambulance: string;
  life_support_ambulance: string;
  house_no: string;
  patient_id: string;
  road_no: string;
  schedule: string;
  sector_no: string;

}